Abstracting demultiplexing phase 1
