+++ taca/__init__.py
@@ -1,4 +1,4 @@
 """ Main TACA module
 """
 
-__version__ = '0.1.7'
+__version__ = '0.1.8'
